Change process design
Now has 1 process per room.  Commands are run through the
CommandDispatcher process, which holds onto the WorldServer process.

Currently, commands can have two scopes: :room and :world.
If the scope is :room, the room process runs the command.  If
the scope is :world, the CommandDispatcher runs it.

This has several deficiencies:

1. It bottlenecks on world commands.  I could add more scope options
than just room and world.

2. Commands require an actor.  I might need a more generic system for
non-actor-related processes that I might have to run in the game.

3. I didn't pay much attention to fault tolerance.

4. Not all player input will need to be run through the room.  Some things
that affect gameplay systems (such as skills) can probably live outside
this design.

5. There's just a singular list of commands that is checked.

But this is good enough for now.